Windows compatibility patch
