+++ pytest.ini
@@ -1,2 +1,3 @@
 [pytest]
 pythonpath = .
+faulthandler_timeout = 5
