+++ tox.ini
@@ -1,5 +1,5 @@
 [tox]
-envlist = py27, py37, py38, py39, py310
+envlist = py27, py37, py38, py39, py311
 skipsdist = True
 
 [testenv]
